sj-blog/learningpython3/requirements.txt

django==4.2.14
django-admin==2.0.2
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-tinymce==4.1.0
pytz==2024.1
psycopg2==2.9.9
screen==1.0.1
xlwt==1.3.0

